chore(reader): fix colors when in iOS 18's dark mode
Fix the UI colors used in BibleReader views when in iOS 18's dark mode.

We have to override lots of default UI colors (some of which are harder to override) because the Reader has a custom Theme which must apply to many views. In these views we should never see differences based on the device's dark/light mode: our Color Theme must override that completely.

ALSO: this upgrades the CI pipeline to the current Xcode and a newer simulator OS release, since github made that change today, and that change is required to test this PR.