feat(reader)!: display footnotes in BibleReaderView
## Description

Displays footnotes in a new sheet of BibleReaderView

For the moment it's not using the standard YouVersion footnote icon; that is expected to change soon.

## Type of Change

- [x] `feat:` New feature (non-breaking change which adds functionality)

## Breaking Changes

- [x] This PR contains **BREAKING CHANGES**

Users of BibleTextView who have implemented the VerseTapAction callback, will need to adjust their callback code to receive three parameters instead of the previous two. 
The new parameters are:
- BibleReference (indicates what verse was tapped), 
- String (ActionType: either "footnote" or "reference", depending what the user tapped upon; see BibleVersionRendering.LinkSchemes),
- [BibleFootnote] (supplies the footnote text in the tapped verse, if any)